Chapter 07/CSV Chat/requirements.txt
langchain
langchain_experimental
streamlit
openai
tiktoken
python-dotenv
pinecone-client
tabulate